// app/types/element.ts
export type Element = {
  number: number
  symbol: string
  name: string
  atomic_mass: number
  category: string
  phase: string
  period: number
  group: number
  xpos: number
  ypos: number
  density: number | null
  melt: number | null
  boil: number | null
  electronegativity_pauling: number | null
  discovered_by: string | null
  block: string
}